release beta version (#8)
* update script

* rollback yarn title

* fix fpioa page

* fix compile

* fix zip on windows

* fix ld args

* fix fpioa

* fix publish script

+++ my-scripts/build-env/codeblocks/removeDir.ts
@@ -1,6 +1,7 @@
 import { rmdir as rmdirAsync, unlink as unlinkAsync } from 'fs';
 import * as rimraf from 'rimraf';
 import { isWin } from '../misc/constants';
+import { isExists } from '../misc/fsUtil';
 import { globalSuccessMessage } from '../misc/globalOutput';
 import { timeout } from '../misc/timeUtil';
 
@@ -14,7 +15,7 @@ function wrapFs(of: Function, output: NodeJS.WritableStream): Function {
 	}) as any;
 }
 
-export function removeDirectory(path: string, output: NodeJS.WritableStream, verbose = true) {
+export async function removeDirectory(path: string, output: NodeJS.WritableStream, verbose = true) {
 	output.write(`removing directory: ${path}...\n`);
 	
 	if (process.cwd().indexOf(path) === 0) {
@@ -22,7 +23,12 @@ export function removeDirectory(path: string, output: NodeJS.WritableStream, ver
 		return Promise.reject(new Error('No way to remove current directory.'));
 	}
 	
-	let p = new Promise<void>((resolve, reject) => {
+	if (!await isExists(path)) {
+		output.write(`never exists...\n`);
+		return;
+	}
+	
+	await new Promise<void>((resolve, reject) => {
 		const wrappedCallback = (err) => err? reject(err) : resolve();
 		
 		rimraf(path, {
@@ -34,19 +40,13 @@ export function removeDirectory(path: string, output: NodeJS.WritableStream, ver
 		}, wrappedCallback);
 	});
 	
-	p = p.then(() => {
-		output.write(`remove complete. delay for OS.\n`);
-	});
+	output.write(`remove complete. delay for OS.\n`);
 	
 	if (isWin) {
-		p = p.then(() => timeout(5000));
+		await timeout(5000);
 	} else {
-		p = p.then(() => timeout(500));
+		await timeout(300);
 	}
 	
-	p = p.then(() => {
-		globalSuccessMessage(`remove directory finish.`);
-	});
-	
-	return p;
+	globalSuccessMessage(`remove directory finish.`);
 }
